refactor(completed-resources): refactor resources controller to align with style guides
